Clean up pre v4 release
