--- bsb/config/parsers/__init__.py
@@ -1,14 +1,0 @@
-from ._parser import Parser
-
-
-def get_parser_classes():
-    from ...plugins import discover
-
-    return discover("config.parsers")
-
-
-def get_parser(parser):
-    return get_parser_classes()[parser]()
-
-
-__all__ = ["Parser", "get_parser", "get_parser_classes"]
