+++ bsb/connectivity/__init__.py
@@ -4,6 +4,5 @@ from .strategy import ConnectionStrategy
 
 # isort: on
 from .detailed import *
-from .detailed.fiber_intersection import FiberTransform, QuiverTransform
 from .general import *
 from .import_ import CsvImportConnectivity
