+++ bsb/connectivity/detailed/__init__.py
@@ -1,3 +1,1 @@
-from .fiber_intersection import FiberIntersection
-from .touch_detection import TouchDetector
 from .voxel_intersection import VoxelIntersection
