+++ bsb/placement/__init__.py
@@ -5,6 +5,5 @@ from .strategy import PlacementStrategy
 # isort: on
 from .arrays import ParallelArrayPlacement
 from .import_ import CsvImportPlacement, ImportPlacement
-from .particle import ParticlePlacement, RandomPlacement
-from .satellite import Satellite
+from .random import RandomPlacement
 from .strategy import Entities, FixedPositions
