+++ docs/getting-started/include_morphos.yaml
@@ -42,13 +42,13 @@ cell_types:
       - neuron_NM
 placement:
   base_placement:
-    strategy: bsb.placement.ParticlePlacement
+    strategy: bsb.placement.RandomPlacement
     cell_types:
     - base_type
     partitions:
     - base_layer
   top_placement:
-    strategy: bsb.placement.ParticlePlacement
+    strategy: bsb.placement.RandomPlacement
     cell_types:
     - top_type
     partitions:
